# Bugfix of MPS from strings and faster QN hashing.